fix env parameter in concept's definition, move getEnvParam function to common ui
